+++ libsolidity/ast/ASTJsonConverter.cpp
@@ -315,13 +315,16 @@ bool ASTJsonConverter::visit(UsingForDirective const& _node)
 	vector<pair<string, Json::Value>> attributes = {
 		make_pair("typeName", _node.typeName() ? toJson(*_node.typeName()) : Json::nullValue)
 	};
+
 	if (_node.usesBraces())
 	{
 		Json::Value functionList;
-		for (auto const& function: _node.functionsOrLibrary())
+		for (auto&& [function, op]: _node.functionsAndOperators())
 		{
 			Json::Value functionNode;
 			functionNode["function"] = toJson(*function);
+			if (op)
+				functionNode["operator"] = string(TokenTraits::toString(*op));
 			functionList.append(move(functionNode));
 		}
 		attributes.emplace_back("functionList", move(functionList));
@@ -811,6 +814,8 @@ bool ASTJsonConverter::visit(UnaryOperation const& _node)
 		make_pair("operator", TokenTraits::toString(_node.getOperator())),
 		make_pair("subExpression", toJson(_node.subExpression()))
 	};
+	if (FunctionDefinition const* function = _node.annotation().userDefinedFunction)
+		attributes.emplace_back("function", nodeId(*function));
 	appendExpressionAttributes(attributes, _node.annotation());
 	setJsonNode(_node, "UnaryOperation", std::move(attributes));
 	return false;
@@ -824,6 +829,8 @@ bool ASTJsonConverter::visit(BinaryOperation const& _node)
 		make_pair("rightExpression", toJson(_node.rightExpression())),
 		make_pair("commonType", typePointerToJson(_node.annotation().commonType)),
 	};
+	if (FunctionDefinition const* function = _node.annotation().userDefinedFunction)
+		attributes.emplace_back("function", nodeId(*function));
 	appendExpressionAttributes(attributes, _node.annotation());
 	setJsonNode(_node, "BinaryOperation", std::move(attributes));
 	return false;